refactor(editor): Extract apiKeys code into features (no-changelog)
