+++ packages/frontend/editor-ui/src/features/apiKeys/apiKeys.constants.ts
@@ -1,0 +1,1 @@
+export const API_KEY_CREATE_OR_EDIT_MODAL_KEY = 'createOrEditApiKey';
