+++ packages/frontend/editor-ui/src/features/apiKeys/components/ApiKeyCard.vue
