+++ packages/frontend/editor-ui/src/features/apiKeys/components/ApiKeyScopes.vue
