+++ packages/frontend/editor-ui/src/features/apiKeys/views/SettingsApiView.vue
@@ -6,19 +6,20 @@ import { useDocumentTitle } from '@/composables/useDocumentTitle';
 
 import { useSettingsStore } from '@/stores/settings.store';
 import { useCloudPlanStore } from '@/stores/cloudPlan.store';
-import { API_KEY_CREATE_OR_EDIT_MODAL_KEY, DOCS_DOMAIN, MODAL_CONFIRM } from '@/constants';
+import { DOCS_DOMAIN, MODAL_CONFIRM } from '@/constants';
+import { API_KEY_CREATE_OR_EDIT_MODAL_KEY } from '../apiKeys.constants';
 import { useI18n } from '@n8n/i18n';
 import { useTelemetry } from '@/composables/useTelemetry';
 import { usePageRedirectionHelper } from '@/composables/usePageRedirectionHelper';
 import { useUIStore } from '@/stores/ui.store';
-import { useApiKeysStore } from '@/stores/apiKeys.store';
+import { useApiKeysStore } from '../apiKeys.store';
 import { storeToRefs } from 'pinia';
 import { useRootStore } from '@n8n/stores/useRootStore';
 
 import { ElCol, ElRow } from 'element-plus';
 import { N8nActionBox, N8nButton, N8nHeading, N8nLink, N8nText } from '@n8n/design-system';
 import { I18nT } from 'vue-i18n';
-import ApiKeyCard from '@/components/ApiKeyCard.vue';
+import ApiKeyCard from '../components/ApiKeyCard.vue';
 
 const settingsStore = useSettingsStore();
 const uiStore = useUIStore();
